[IMP] website_shift: Next shifts regular worker
Add configuration to specify how many next shifts for regular worker
must be shown to them on their personal web page. This config is called
`regular_next_shift_limit`.

Improve the view that shows the next shifts on the personal page for a
regular worker regarding to the `regular_next_shift_limit`.

Because all the next shifts are not created in the database, we need to
create 'fictive' one that are not stored.