[FIX] webs_shift: Wrong display of date last shift
Fix the case when an irregular worker is in alert and that the date
before last shift is displayed.

The `date_before_last_shift` is renamed in `future_alert_start_time`.
This name better represent what this variable contains. It also explains
why this information does not have to be displayed when the worker is
already in alert mode as this date is the date in the future on which
the worker will be in alert.