[ADD] website_shift: Personal page for working information (#33)
* [ADD] website_shift: Add configuration

irregular_shift_limit: Maximum shift that will be shown
highlight_rule: Treshold of available space in a shift that trigger the
                highlight of the shift
hide_rule: Treshold ((available space)/(max space)) in percentage of
           available space under wich the shift is hidden
The last one replaces the old hide filter.

* [FIX] website_shift: Wrong mail on regular shift

Fix the wrong email address that is shown on the public page that show
available shift for irregular worker.

* [ADD] website_shift: Irregular personal work page

Add personal shift page for irregular worker that allows you to
subscribe to a shift.

* [ADD] website_shift: Personal page for non-worker

Add a personal page for non-worker that shows a simple message telling
them that they don't have to do a shift.

* [ADD] website_shift: Regular personal work page

Add personal shift page for regular worker.

* [ADD] website_shift: Exempted personal work page

Add personal shift page for exempted worker.

* [IMP] website_shift: Shopping status personal page

This add a label that shows the shopping status of a user.
If the label is green the user can shop.
If the label is red the user cannot shop.
There is also icons for disambiguation.

* [ADD] website_shift: Config en & dis shift sign up

This add a configuration boolean that enable or disable the shift sign
up for irregular worker.

* [IMP] website_shift: Status on personal pages

Show dates properly.
Hide what's useless.
Give meaningful name.

* [IMP] website_shift: Design of personal pages

* Lower size of h2 title.
* Add top margin to h1.
* Change highlight color.
* Change position of explanation test.

* [IMP] website_shift: Refactor templates

Refactor templates to avoid dupplicate code.
Refactor the controller to match with new templates.
Add license.

* [ADD] website_shift: Past shifts on personal page

Add a list of the previous shifts for a worker.

* [ADD] website_shift: Super Coop info

Add a button to show email and phone of the super cooperator of a shift.
Tell you if you are a Super Co-operator

* [IMP] website_shift: Next shifts regular worker

Add configuration to specify how many next shifts for regular worker
must be shown to them on their personal web page. This config is called
`regular_next_shift_limit`.

Improve the view that shows the next shifts on the personal page for a
regular worker regarding to the `regular_next_shift_limit`.

Because all the next shifts are not created in the database, we need to
create 'fictive' one that are not stored.

* [ADD] website_shift: Warning about holidays

Add a warning message on the personal page for a regular worker that
tells that public holidays are not taken into account for the moment.

* [IMP] website_theme: colors for warnings

* [FIX] website_shift: 403 when subscribing

Fix the case when an irregular worker that doesn't have write access to
beesdoo.shift.shift try to subscribe to a shift and got a 403 error.

* [FIX] website_theme: Theme must be an application

As wrote in this document :
https://www.odoo.com/documentation/9.0/howtos/themes.html#create-a-theme-module
a theme module must be an application.

* [FIX] webs_shift: Wrong display of date last shift

Fix the case when an irregular worker is in alert and that the date
before last shift is displayed.

The `date_before_last_shift` is renamed in `future_alert_start_time`.
This name better represent what this variable contains. It also explains
why this information does not have to be displayed when the worker is
already in alert mode as this date is the date in the future on which
the worker will be in alert.

* [FIX] website_shift: Co-operator -> Cooperator

* [FIX] website_shift: Past shifts not ordered desc

The past shift was not properly ordered. This fix by ordering the past
shifts from the newest to the oldest.

* [FIX] website_shift: Holiday shown after end of it

Before:
    The holiday is shown even if the holiday is in the past.
After:
    The holiday is shown only if the holiday is now or in the future.