[FIX] website_shift: Error when regular has no tz
Issue when no timezone set to a regular worker. It try to use the user
timezone but has it doesn't exists runs into an error.
To solve this, it first takes the timezone of the user, if it is
empty, it takes the timezone of the context, and if also empty it uses
UTC.