[FIX] b_w_shift: Issue with settings
There was an issue that shows an error message when trying to modify the
settings for shift in the website configuration.

This is due to a missing sequence in the database. I don't know why this
sequence where not present. But a simple query create it if it doesn't
exists.