[CHG] b_shift: Remove alert_start_time when not in alert status
Before this commit, the alert_start_time was not erased when a status
change from 'alert' to 'ok' for an irregular worker.

People find that unintuitive (despite this date was in the past when
you are in 'ok' status).

Now, this alert_start_time was erased when the status change from
'alert' to 'ok'.