[FIX] b_w_shift: Shift with no start and end time
When the worker has subscribed to a shift template but not yet to a
shift, the shift used to generate her next shift is the first shift
found in the shift template (so this is a shift that is attached to
another worker from the same shift template).

Now the system pays attention to not use a shift that has no start time
or end time defined.