[CHG] b_base: Remove cooperator_type info
This new b_base module focus on adding member cards and eater
functionality.

Constrain about number of eater should be customise for each
cooperative. And if the cooperative is using easy_my_coop module the
maximum number of eater should be specified on the share product.

Also cooperator_type has no meaning here. For cooperative that uses
easy_my_coop, cooperator_type is a computed field based on the share
that the cooperator owns.

beesdoo_easy_my_coop and beesdoo_worker_status will be modified for
taking this into account.