[FIX] _compute_fake_date trigger
In odoo 16, computed fields are now stored in cache. This means that the start_date and end_date field, are not recomputed when running _generate_next_planning multiple times with a different visualize_date in context.
This fix add a dependency in _compute_fake_date on the context, which solves the problem.