[FIX] beesdoo_shift_swap: Prevent a skip of subsequent items
Modifying a list while looping over it is not allowed. It causes
subsequent items to be skipped.

I suspect that this is the cause of a bug where a person who wants to
swap a shift is correctly unsubscribed from their old shift, but not
subscribed to their new shift. But I have not been able to verify this.

Signed-off-by: Carmen Bianca BAKKER <carmen@coopiteasy.be>