[FIX] pos_shift_partner_can_shop: popup does not show
Using actionName is not correct because it is subject to languages
changes. If language is set to something else than English then this
test will fail.

A side effect, is that clicking of the "Select partner" button does not
distinguish from clicking of the "Payment" button with such a test.

The right way is to get event_name that comes as the first argument of
the `trigger` method.